Fix bug with NPE if by some reason tokens are absent in db but GUI sends it

+++ delis-web-api/src/main/java/dk/erst/delis/rest/controller/logout/RestLogoutController.java
@@ -33,7 +33,9 @@ public class RestLogoutController {
     public ResponseEntity<DataContainer<Boolean>> logout(HttpSession session) {
         OAuth2Authentication authentication = (OAuth2Authentication) SecurityUtil.getAuthentication();
         OAuth2AccessToken token = ((DefaultTokenServices) consumerTokenServices).getAccessToken(authentication);
-        tokenService.removeAccessToken(token);
+        if (token != null) {
+        	tokenService.removeAccessToken(token);
+        }
         session.invalidate();
         return ResponseEntity.ok(new DataContainer<>(true));
     }
